- Introducing user safety Now we have a queue of users being processed. This permits to only accept repositories belonging to the users being processed.
On one side this brings added safety but the real reason was to permit different clients to process the same user. This is useful when a given client takes too long to process a user (for example waiting for the rate limit to expire) or when the client suddenly goes offline without notice. This way we don't lose the user, it will take a bit longer to process (2 hours waiting period).

Also, we now don't accept duplicate repositories. This means that the first client to submit a repository will not get overwritten. This procedure solves the concurrency issues when two clients are simultaneously processing the repositories from the same user.

The server/client is reaching a state of robustness and capacity to resume work in faulty situations. It is currently missing to account with the situation of the server failing and losing the users/repositories that are currently being processed. Will later implement the support for storing this information on a flat file and to recover when starting the program from scratch.

At this point no security measures were added, we trust clients to provide reliable information. This will be improved as development moves forward.