- Adding save/load state when the server is terminated abruptly
This new feature will be saving to disk the list of users currently being processed. The big advantage is that whenever the server is turned off we don't lose the information about users that were on the queue. The restored list of users will be placed with priority for processing, avoiding that these users get too far from their original position in the index.